Player Collisions and facing direction
Implement player collisions using the arcade physics system and the collide method. In order to now pass on the players velocity the immovable parameter on gameworld assets has been set to true.
The Players facing direction has been logged using the keys list in preperation for the future.
The space action key has been refactored out of the keys list to keep list and the detect movement function focused on movement specific functionality.