Template Race Data
Create an array of Races which each have a name and an array of levels.
Use that array to loop and draw out buttons for each race on the race select scene and each level on the race level select scene.

Set and pass the active race and active level while moving through the race scenes.
Display the race type and level on the race scene.