RaceScene
Add text to display pet position.
Refactor update logic to stop all pets on the finish line and mark the race as ended when all pets have reached the end.
add a new tile type for the finish line and a new finish overlap function that plays the celebrate animation and sets gravity and velocity to 0
track the main pet position and display it with text in the top corner.
record pet positions as they reach the finish line so the top 3 can be displayed in the results scene.