Hitboxes and racing tweaks
Reduce the size of the egg image to fix hitbox scaling bugs,
Add new tiles for ground, water and climbing.
add a new test map.
Adjust tile offset values and set the hitbox sizes of tiles to make pets overlap near the center of the tile.
Bring the active pet to the front lane and adjust the camera bounds to fit all the lanes in nicely on screen.