Race Refactor
Refactor the race to use just a single lane and have all pets overlap with the same lane. To give the illusion that they are running on differnt paths each pet has their physics body offset slightly. Using a single lane of physics tiles reduces the number of tile phyics bodies by 8*.

Refactor the tile creation to a method createTile() to reduce repeated code.
Add a method createPet() to reduce the size of the create method.
Add a start tile group to place each pet offset to the starting race tile.